Guard against user fast-forwarding over mid-roll
Due to the combination of calling player.play() when the renderer
has no ads and IMACuepoint returning isPlayed==true as the user
fast forwarded through the mid-roll, the player allowed users
to seek past a mid-roll if no true[X] ads were available.

With this change, we detect when a seek is reported to occur
with a start time within a mid-roll (meaning the viewer started
seeking before the mid-roll), and resume playback at the first
video ad in the ad pod.