Improve view transitions between home and player views
Mute the audio on the home screen preview. Also pause the
home screen preview video when we transition to the player.

Correct the behavior of pressing the "MENU" button in the
stream to exit back to the home screen on a single press.

Fix behavior of the "MENU" button for the user cancel
stream event.